Use bag split as signal from recorder in rosbag tests
The rosbag tests are flaky, because it can happen that the recorder is
stopped before it has the required number of messages received. This is
because the process that contains the recorder has a second subscriber
that keeps track of the number of messages published. However, the
recorder's subscriber and that counting subscriber are not synchronized.
Even if the message counting subscriber has received some number of
messages, this doesn't mean that the recorder's subscriber has received
even a single message.

We fix the issue by replacing the message counting subscriber by one
that listens for bag split events. We also configure the recorder to
split bags once some number of bytes has been recorded. Once the bag
split event is received, we can be certain that the recorder has
received that amount of bytes.