Add patch to move ts_lib to PublisherBase
This adds a patch that moves the `std::shared_ptr` holding a reference
to a `rclcpp::SharedLibrary` from `rclcpp::GenericPublisher` to
`rclcpp::PublisherBase`. The shared library in question contains type
support code for the publisher's message type. In the standard ROS
build, two separate shared objects are used, one is represented by the
mentioned `shared_ptr`-wrapped `SharedLibrary` instance, the other is
never unloaded. At the same time, code from the latter library is
required during destruction of a `PublisherBase` instance. The
rules_ros2 build creates a single library, which is unloaded when
`GenericPublisher` is destroyed (the `shared_ptr`'s reference count goes
to zero). When `GenericPublisher`'s parent class, `PublisherBase`, is
destroyed, trying to call code from the shared library failed and caused
a segfault. Moving the `shared_ptr` to `PublisherBase` makes sure the
shared library stays loaded as long as required.

We further extend the existing rosbag test to also excersise the `play`
command.

This fix is similar to what was done for PR#47. See the discussion there
for more details.

Fixes #110